feat: implement parser for artworks
Implementation rational:

Artworks is part of the Google Knowledge Graph. these elements are tagged  with KC since 2019, so it is quite reliable. Even though in 2019 “Artworks” was not included in the group tagged with KC, in 2024 and today it is—so we can expect it to remain moving forward.

The Google parser will iterate over all KC entries. If the method is implemented, it will be added to the final hash; if not, it will log the result so we can implement it later.

The entry class is Search. The executeQuery method is not yet implemented, but this is the part that would actually fetch the data from the Google API.

I prefer to start with end-to-end testing to ensure that the features work. This way, I can change the internal code without having to rewrite tests, as I am testing the features rather than how the code is written. The downside is that it will over-test some lines of code—for example, search functions are called repeatedly. The benefit is that we know the feature works end-to-end, including all plumbing. Otherwise, we would have to write additional tests to ensure the plumbing works.

I have ensured that i use tags/attributes that are part of the accessability tree for long term reliablity as companies like google are required to implement this properly by law, as opposed to using css classes as these can change any time

My implementation of extensions may not be adaquete for all cases, but its a good start. an example would be if it has the painting name date and artist i tried to find such a case by searching "post impressionism artworks" but the extension would be the artist name only

Alternative:

While i have optimised for long term stability and extensability, there are potential performance issues. if this where the case we could just query by css tags and use one regex query with index lookups for image data. and if we still have issues we would just build regex by hand instead of using nokogiri